% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{mf}
\alias{mf}
\alias{mf,Consensus-method}
\alias{adductRules}
\alias{adductRules,Consensus-method}
\alias{organism}
\alias{organism,Consensus-method}
\alias{database}
\alias{database,Consensus-method}
\alias{threshold}
\alias{threshold,Consensus-method}
\alias{hits}
\alias{hits,Consensus-method}
\alias{PIPs}
\alias{PIPs,Consensus-method}
\alias{classifications}
\alias{classifications,Consensus-method}
\alias{consensusClassifications}
\alias{consensusClassifications,Consensus-method}
\title{Accessor methods for the \code{Consensus} S4 class}
\usage{
mf(x)

\S4method{mf}{Consensus}(x)

adductRules(x)

\S4method{adductRules}{Consensus}(x)

organism(x)

\S4method{organism}{Consensus}(x)

database(x)

\S4method{database}{Consensus}(x)

threshold(x)

\S4method{threshold}{Consensus}(x)

hits(x)

\S4method{hits}{Consensus}(x)

PIPs(x)

\S4method{PIPs}{Consensus}(x)

classifications(x)

\S4method{classifications}{Consensus}(x)

consensusClassifications(x)

\S4method{consensusClassifications}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class \code{Consensus}}
}
\description{
Accessor methods for the \code{Consensus} S4 class.
}
