% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{adductRules}
\alias{adductRules}
\alias{adductRules,Consensus-method}
\title{adductRules}
\usage{
adductRules(x)

\S4method{adductRules}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the adduct rules of a Consensus object.
}
