% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{as_por}
\alias{as_por}
\title{Convert matrix to profile of rankings}
\usage{
as_por(matrix, criterion = NULL)
}
\arguments{
\item{matrix}{Matrix to convert to a profile of rankings}
}
\value{
A profile of rankings object
}
\description{
Convert matrix to profile of rankings
}
