% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{borda}
\alias{borda}
\title{Borda Count ranking rule}
\usage{
borda(profileOfRankings, seeTrace = FALSE, seePoints = FALSE)
}
\arguments{
\item{seeTrace}{}
}
\value{

}
\description{
Apply Borda Count ranking rule in a profile of ranking
}
\seealso{
Other ranking rule: 
\code{\link{borda_winner}()},
\code{\link{scoring_rule}()},
\code{\link{tapproval}()},
\code{\link{veto}()}
}
\concept{ranking rule}
