% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R, R/scoringc.R
\name{borda_count}
\alias{borda_count}
\title{Apply Borda Count ranking rule in a profile of ranking}
\usage{
borda_count(profileOfRankings, verbose = FALSE, seePoints = FALSE)

borda_count(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{



}
\description{
Apply Borda Count ranking rule in a profile of ranking

Title
}
\seealso{
Other ranking rule: 
\code{\link{borda_winner}()},
\code{\link{scoring}()},
\code{\link{tapproval}()},
\code{\link{veto}()}
}
\concept{ranking rule}
