% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ranking_rule_c.R
\name{borda_winner}
\alias{borda_winner}
\title{Title}
\usage{
borda_winner(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
