% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring_points.R
\name{calculatePoints}
\alias{calculatePoints}
\title{Calculate points for scoring ranking ruless}
\usage{
calculatePoints(ranking, method = NULL, t = 0, seeTrace = F, seePoints = F)
}
\arguments{
\item{ranking}{}

\item{seePoints}{}
}
\description{
Calculate points for scoring ranking ruless
}
