% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{classifications}
\alias{classifications}
\alias{classifications,Consensus-method}
\title{classifications}
\usage{
classifications(x)

\S4method{classifications}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the classifications of a Consensus object.
}
