% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{condorcet_loser}
\alias{condorcet_loser}
\title{Condorcet loser}
\usage{
condorcet_loser(profileOfRankings)
}
\arguments{
\item{profileOfRankings}{}
}
\value{

}
\description{
A candidate is preferred by less or equal than half of the voters to all the other 
candidates, then it is presumptively the worst and then it is selected as
Condorcet loser
}
