% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{condorcet_winner}
\alias{condorcet_winner}
\title{Condorcet winner}
\usage{
condorcet_winner(profileOfRankings, weak = FALSE)
}
\arguments{
\item{profileOfRankings}{}
}
\value{

}
\description{
A candidate is preferred by more than half of the voters to all the other 
candidates, then it is presumptively the best and then it is selected as
Condorcet winner
}
