% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Consensus-class}
\alias{Consensus-class}
\title{Consensus}
\description{
An S4 class to store consensus structral classification results for a molecular formula.
}
\section{Slots}{

\describe{
\item{\code{MF}}{molecular formula}

\item{\code{adductRules}}{tibble containing adduct rules as returned by mzAnnotation::adducts()}

\item{\code{organism}}{organism KEGG ID. NA if database is pubchem.}

\item{\code{database}}{database, kegg or pubchem}

\item{\code{threshold}}{% majority for consensus classification}

\item{\code{hits}}{MetaboliteDatabase containing molecular formula hits from given database}

\item{\code{PIPs}}{putative ionisation product matches from database hits for given adduct rules}

\item{\code{classifications}}{structural classifications for database hits}

\item{\code{consensus}}{consensus structural classifications for adducts}
}}

