% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensusClassification}
\alias{consensusClassification}
\title{consensusClassification}
\usage{
consensusClassification(MF, adducts = c("[M-H]1-"), threshold = 0.5)
}
\arguments{
\item{MF}{molecular formula}

\item{adducts}{character vector of adducts}

\item{threshold}{consensus threshold}
}
\description{
Calculate a consensus classification for a given molecular formula and adducts.
}
