% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{consensusClassifications}
\alias{consensusClassifications}
\alias{consensusClassifications,Consensus-method}
\title{consensusClassifications}
\usage{
consensusClassifications(x)

\S4method{consensusClassifications}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the consensus classifications of a Consensus object.
}
