% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Consensuses-class}
\alias{Consensuses-class}
\title{Consensuses}
\description{
An S4 class to store consensus structural classifications.
}
\section{Slots}{

\describe{
\item{\code{consensuses}}{list containing consensus results for each MF}

\item{\code{results}}{tibble containing consensus structural classifications for each feature}
}}

