% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/consensus.R
\docType{methods}
\name{construct}
\alias{construct}
\alias{construct,Assignment-method}
\title{construct}
\usage{
construct(x, organism = "hsa", threshold = 0.5, databases = c("kegg",
  "pubchem"))

\S4method{construct}{Assignment}(x, organism = "hsa", threshold = 0.5,
  databases = c("kegg", "pubchem"))
}
\arguments{
\item{x}{S4 object of class Workflow}

\item{organism}{organism kegg ID.}

\item{threshold}{majority assignment threshold for consensus classifications}
}
\description{
Consensus structural classifications for molecular formula assignments.
}
\examples{
\dontrun{
library(MFassign) 
p <- assignmentParameters('FIE')
p@nCores <- 2
assignment <- assignMFs(peakData,p)

consensusCl <- construct(assignment)
}
}
