% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{construction}
\alias{construction}
\title{construction}
\usage{
construction(MFs, path = ".", db = c("kegg", "pubchem"),
  organism = character(), adductRules = adducts(), threshold = 50)
}
\arguments{
\item{MFs}{vector of molecular formulas}

\item{path}{target file path for classification library}

\item{db}{databases to search. Can be either kegg or pubchem.}

\item{organism}{KEGG organism ID. Ignored if kegg is not specified in db.}

\item{adductRules}{adduct rules table as returned by mzAnnotation::adducts()}

\item{threshold}{\% majority threshold for consensus classifications}
}
\description{
Build or add to a consensus classification library.
}
