% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{database}
\alias{database}
\alias{database,Consensus-method}
\title{database}
\usage{
database(x)

\S4method{database}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the database of a Consensus object.
}
