% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{get_ranking}
\alias{get_ranking}
\title{Get a single ranking from a profile of rankings}
\usage{
get_ranking(profileOfRankings, index)
}
\arguments{
\item{profileOfRankings}{object to get the ranking}

\item{index}{row of the ranking to extract}
}
\description{
The ranking is not deleted from the profile.
}
