% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{hits}
\alias{hits}
\alias{hits,Consensus-method}
\title{hits}
\usage{
hits(x)

\S4method{hits}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the database molecular formula matches of a Consensus object.
}
