\name{Huex}
\alias{Huex}
\docType{data}
\title{
Affymetrix Huex gene expression data
}
\description{
Gene expression data from 27 Glioblastoma Multiforme (GBM) patients measured on the Affymetrix HuEx GeneChip.
}
\usage{data("TCGA")}
\format{
Numeric matrix.
}
\source{
\url{https://tcga-data.nci.nih.gov/docs/publications/gbm_exp/LBL202.txt}
}
\references{
Verhaak, R. G. W., Hoadley, K. A., Purdom, E., Wang, V., Qi, Y., Wilkerson, M. D., ..., Cancer Genome Atlas Research Network. Integrated genomic analysis identifies clinically relevant subtypes of glioblastoma characterized by abnormalities in PDGFRA, IDH1, EGFR, and NF1. \emph{Cancer Cell}, 2010, 17(1), 98-110. 
}
\examples{
data("TCGA")
}
\keyword{datasets}
