% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_irv.R
\name{irv}
\alias{irv}
\title{Instant runoff voting}
\usage{
irv(profileOfRankings, break_ties = "random", seeTrace = FALSE)
}
\arguments{
\item{profileOfRankings}{}
}
\value{

}
\description{
Instant runoff voting
}
