% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{is.ranking}
\alias{is.ranking}
\title{Check if the object is a ranking}
\usage{
is.ranking(x, ...)
}
\arguments{
\item{x}{Object to check whether is a ranking}

\item{...}{Any other parameter will be ignored}
}
\value{
TRUE if the object given as parameter is a ranking. FALSE otherwise.
}
\description{
Check if the object is a ranking
}
