% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{keggCompoundInfo}
\alias{keggCompoundInfo}
\title{keggCompoundInfo}
\usage{
keggCompoundInfo(IDs)
}
\arguments{
\item{IDs}{a vector of KEGG compound entry IDs}
}
\description{
Retrieve KEGG compound information.
}
\examples{
keggCompoundInfo(c('C00089','C00149'))
}
