% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{keggCompounds}
\alias{keggCompounds}
\title{keggCompounds}
\usage{
keggCompounds(organism = NULL)
}
\arguments{
\item{organism}{KEGG organism ID. If NULL use all compounds.}
}
\description{
Return KEGG compound accession IDs for a given organism
}
\examples{
\dontrun{
keggCompounds('hsa')
}
}
