% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/kegg.R
\docType{methods}
\name{keggConsensus}
\alias{keggConsensus}
\alias{keggConsensus,Assignment-method}
\title{keggConsensus}
\usage{
keggConsensus(x, organism = "hsa", threshold = 0.5)

\S4method{keggConsensus}{Assignment}(x, organism = "hsa",
  threshold = 0.5)
}
\arguments{
\item{x}{S4 object of class Assignment}

\item{organism}{organism kegg ID.}

\item{threshold}{majority assignment threshold for consensus classifications}
}
\description{
Collate consensus classifications for molecular formula assignments using KEGG.
}
