% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/kegg.R
\docType{methods}
\name{keggConstruct}
\alias{keggConstruct}
\alias{keggConstruct,Assignment-method}
\title{keggConstruct}
\usage{
keggConstruct(x, organism = "hsa", threshold = 0.5)

\S4method{keggConstruct}{Assignment}(x, organism = NULL,
  threshold = 0.5)
}
\arguments{
\item{x}{S4 object of class Assignment}

\item{organism}{organism kegg ID. If NULL use all compounds.}

\item{threshold}{majority assignment threshold for consensus classifications}
}
\description{
Collate consensus classifications for molecular formula assignments using KEGG.
}
