% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{keggPIPs}
\alias{keggPIPs}
\title{keggPIPs}
\usage{
keggPIPs(MFs, organism = NULL, adductRules = adducts())
}
\arguments{
\item{MFs}{tibble containing 2 columns. The first named MF containing molecular formulas. The second named Adduct containing adducts. See example.}

\item{organism}{KEGG organism ID. If NULL use all compounds.}
}
\description{
Return KEGG putative ionisation products (PIPs) for given molecular formulas and their adducts.
}
\examples{
\dontrun{
d <- tibble(MF = c('C12H22O11','C4H6O5'),Adduct = rep('[M-H]1-',2))
keggPIPs(d,'hsa')
}
}
