% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{kemeny_distance}
\alias{kemeny_distance}
\title{Kemeny distance between two rankings}
\usage{
kemeny_distance(r1, r2, seeTrace = FALSE)
}
\arguments{
\item{r1}{}

\item{r2}{}
}
\value{

}
\description{
Kemeny distance between two rankings
}
