% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_kendall.R
\name{kendall_ranking_to_por}
\alias{kendall_ranking_to_por}
\title{Kendall distance in a profile of rankings}
\usage{
kendall_ranking_to_por(profileOfRankings, ranking, verbose = FALSE)
}
\arguments{
\item{verbose}{}

\item{ranking1}{}

\item{ranking2}{}
}
\value{

}
\description{
Kendall distance in a profile of rankings
}
