% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{kendallpor}
\alias{kendallpor}
\title{Kendall distance in a profile of rankings}
\usage{
kendallpor(profileOfRankings, verbose = FALSE)
}
\arguments{
\item{verbose}{}

\item{ranking1}{}

\item{ranking2}{}
}
\value{

}
\description{
Kendall distance in a profile of rankings
}
