% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_perm_litvak.R
\name{litvak}
\alias{litvak}
\title{Litvak ranking rule}
\usage{
litvak(profileOfRankings, parallel = FALSE, seeTrace = FALSE)
}
\arguments{
\item{profileOfRankings}{an object of the class "por"}

\item{seeTrace}{}
}
\value{
litvak returns an object of class litvak
}
\description{
Litvak ranking rule
}
\examples{
por <- parse_profile_of_rankings("6, a ≻ b ≻ c ≻ d,
                                  5, b ≻ c ≻ a ≻ d,
                                  3, c ≻ d ≻ a ≻ b")
litvak(por)
}
