% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{litvak_distance}
\alias{litvak_distance}
\title{Litvak distance between two rankings}
\usage{
litvak_distance(r1, r2, seeTrace = FALSE)
}
\arguments{
\item{r1}{}

\item{r2}{}
}
\value{

}
\description{
Litvak distance between two rankings
}
