% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{Majority}
\alias{Majority}
\alias{MajorityRule}
\alias{MR}
\title{Majority-rule consensus tree}
\usage{
Majority(trees, p = 0.5)

MajorityRule(trees, p = 0.5)

MR(trees, p = 0.5)
}
\arguments{
\item{trees}{A list of trees, or a \code{multiPhylo} object; all entries must
share the same leaf labels.}

\item{p}{Numeric between 0.5 and 1: the minimum proportion of trees that must
contain a split for it to be retained.  \code{p = 0.5} (the default) gives the
majority-rule consensus; \code{p = 1} gives the strict consensus.}
}
\value{
\code{Majority()} returns the consensus tree, an object of class \code{phylo},
rooted as in the first entry of \code{trees}.
}
\description{
\code{Majority()} returns the majority-rule consensus
\insertCite{MargushMcMorris1981}{ConsTree}: the tree containing each split
that occurs in more than half of the input trees.  Raising \code{p} retains only
splits present in a greater proportion of trees, up to the strict consensus
at \code{p = 1}.
}
\details{
The majority-rule consensus belongs to the family of consensus methods for
which Jansson and colleagues developed asymptotically efficient algorithms
\insertCite{JanssonShenSung2016}{ConsTree} -- the algorithmic backbone of this
package -- and which their FACT toolkit implements.  This particular method is
a thin wrapper around \code{\link[TreeTools:Consensus]{TreeTools::Consensus()}}, to which the computation is
delegated.
}
\examples{
trees <- ape::as.phylo(0:5, 8)
Majority(trees)

}
\references{
\insertAllCited{}
}
\seealso{
Other consensus methods:
\code{\link[=Adams]{Adams()}},
\code{\link[=Average]{Average()}},
\code{\link[=Frequency]{Frequency()}},
\code{\link[=Greedy]{Greedy()}},
\code{\link[=Local]{Local()}},
\code{\link[=Loose]{Loose()}},
\code{\link[=MajorityPlus]{MajorityPlus()}},
\code{\link[=Quartet]{Quartet()}},
\code{\link[=RStar]{RStar()}},
\code{\link[=Strict]{Strict()}}
}
\concept{consensus methods}
