% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{mf}
\alias{mf}
\alias{mf,Consensus-method}
\title{mf}
\usage{
mf(x)

\S4method{mf}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get and the molecular formula of a Consensus object.
}
