% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{organism}
\alias{organism}
\alias{organism,Consensus-method}
\title{organism}
\usage{
organism(x)

\S4method{organism}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the organism ID of a Consensus object.
}
