% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{parse_profile_of_rankings}
\alias{parse_profile_of_rankings}
\title{Parse profile of rankings}
\usage{
parse_profile_of_rankings(string)
}
\arguments{
\item{string}{A string containing a representation of the ranking in the form:
"number_of_voters_r1, ranking1,
 number_of_voters_r2, ranking2,
 ...,
 number_of_voters_rN, rankingN"
 The operators used for representing that one candidate is preferred to 
 another are ≻ and >. For representing ties use ~.}
}
\value{
A profile of rankings object containing the representation.
}
\description{
Create profile of rankings using its string representation.
}
\examples{

parse_profile_of_rankings("6, a ≻ b ≻ c ≻ d,
                           5, b ≻ c ≻ a ≻ d,
                           3, c ≻ d ≻ a ≻ b")
}
