% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{PIPs}
\alias{PIPs}
\alias{PIPs,Consensus-method}
\title{PIPs}
\usage{
PIPs(x)

\S4method{PIPs}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the putative ionisation products of a Consensus object.
}
