% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSankey}
\alias{plotSankey}
\alias{plotSankey,tbl_df-method}
\alias{plotSankey,Construction-method}
\title{Sankey plot}
\usage{
plotSankey(
  x,
  exclude_kingdoms = c("Unclassified", "No database hits", "Unknown")
)

\S4method{plotSankey}{tbl_df}(
  x,
  exclude_kingdoms = c("Unclassified", "No database hits", "Unknown")
)

\S4method{plotSankey}{Construction}(
  x,
  exclude_kingdoms = c("Unclassified", "No database hits", "Unknown")
)
}
\arguments{
\item{x}{The consensus structural classifications to plot. This should either a tibble of an object of S4 class \code{Construction} as returned by \code{construction()}.}

\item{exclude_kingdoms}{a vector of kingdoms names to exclude from the plot}
}
\description{
Plot an overview of consensus structural classifcations using a sankey plot.
}
\examples{
plotSankey(construction::structural_classifications)
}
