% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ranking_rule.R,
%   R/scoring_ranking_rule_c.R
\name{plurality}
\alias{plurality}
\title{Title}
\usage{
plurality(profileOfRankings, verbose = FALSE, seePoints = FALSE)

plurality(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{



}
\description{
Title

Title
}
