% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pubchemConsensus}
\alias{pubchemConsensus}
\title{pubchemConsensus}
\usage{
pubchemConsensus(MF, adducts = c("[M-H]1-"), threshold = 0.5,
  adductRules = adducts())
}
\arguments{
\item{MF}{molecular formula}

\item{adducts}{character vector of adducts}

\item{threshold}{consensus threshold}

\item{adductRules}{Adduct formation rules to use for putative ionisation products. Defaults to \code{mzAnnotation::adducts()}}
}
\description{
Calculate a consensus classification for a given molecular formula and adducts based on the pubchem database.
}
\examples{
\dontrun{
pubchemConsensus('C10H10O7')
}
}
