% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pubchemMatch}
\alias{pubchemMatch}
\title{pubchemMatch}
\usage{
pubchemMatch(MF)
}
\arguments{
\item{MF}{a molecular formula to query}
}
\description{
Query the pubchem database for molecular formula matches.
}
\examples{
\dontrun{ 
pubchemMatch('C12H22O11')
}
}
