% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\docType{methods}
\name{pubchemPIPs,Consensus-method}
\alias{pubchemPIPs,Consensus-method}
\title{pubchemPIPs}
\usage{
\S4method{pubchemPIPs}{Consensus}(ips)
}
\description{
pubchemPIPs
}
\examples{
pubchemPIPs('C4H6O5', '[M-H]1-')
}
