% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{random_profile_of_rankings}
\alias{random_profile_of_rankings}
\title{Create random profile of rankings}
\usage{
random_profile_of_rankings(
  ncandidates = 4,
  nvoters = 10,
  seed = NULL,
  withties = FALSE,
  cnames = NULL,
  distinct = NULL
)
}
\arguments{
\item{ncandidates}{Number of candidates}

\item{seed}{Fix a seed to hace reproducible code}

\item{withties}{The rankings are generated without ties by default.}

\item{nranking}{Number of rankings in the profile of rankings}
}
\description{
Creates a ranfom profile of rankings
}
