% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Default behaviour creates an ascending value. This is, the lower the value
the better the candidate is considered
If names are given in n it will overwrite the names in the vector if named}
\usage{
ranking(v, cnames = NULL, decreasing = FALSE)
}
\description{
Default behaviour creates an ascending value. This is, the lower the value
the better the candidate is considered
If names are given in n it will overwrite the names in the vector if named
}
