% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking_has_ties}
\alias{ranking_has_ties}
\title{Ranking has ties}
\usage{
ranking_has_ties(ranking)
}
\arguments{
\item{ranking}{Object of class ranking.}
}
\value{
TRUE if the ranking has ties, FALSE otherwise.
}
\description{
Check whether the ranking object given as parameter has ties.
}
