\name{RNASeq}
\alias{RNASeq}
\docType{data}
\title{
RNA-Seq gene expression data
}
\description{
Limma-voom normalised gene expression data from 27 Glioblastoma Multiforme (GBM) patients measured via RNA-Seq.
}
\usage{data("TCGA")}
\format{
Numeric matrix.
}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\references{
Verhaak, R. G. W., Hoadley, K. A., Purdom, E., Wang, V., Qi, Y., Wilkerson, M. D., ..., Cancer Genome Atlas Research Network. Integrated genomic analysis identifies clinically relevant subtypes of glioblastoma characterized by abnormalities in PDGFRA, IDH1, EGFR, and NF1. \emph{Cancer Cell}, 2010, 17(1), 98-110. 
}
\examples{
data("TCGA")
}
\keyword{datasets}
