% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ranking_rule.R
\name{scoring}
\alias{scoring}
\title{Scoring ranking rules}
\usage{
scoring(profileOfRankings, method = NULL, ones = 1, verbose = FALSE)
}
\arguments{
\item{profileOfRankings}{object of the class profile of rankings}

\item{method}{\itemize{
 \item{"plurality"}{Stuff}
 \item{"borda"}{Stuff}
 \item{"ones"}{Stuff}
}}

\item{ones}{only necessary when the \code{method} choosen is}

\item{verbose}{by default FALSE. Change to TURN for seeing on the screen
the workflow of the function}
}
\value{

}
\description{
Scoring ranking rules
}
