% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_simpson.R
\name{simpson}
\alias{simpson}
\title{Simpson}
\usage{
simpson(profileOfRankings)
}
\arguments{
\item{profileOfRankings}{}
}
\value{

}
\description{
Simpson ranks candidates according to their maximum defeat. This means that 
the best ranked candidates are the ones that are not clearly defeated by any 
other candidate.
}
\details{
Using the votrix, for each candidate it is necessary to seek for the value 
of the row that represents the candidate which is the highest of those that 
are lower than half of the voters of the profile of rankings. Then, the 
winning ranking is obtained by ordering the candidates by increasingly value 
of the biggest defeat.
Simpson ranking rule is also a Condorcet method. As it ranks candidates 
according to their maximum defeat, the best ranked candidates are the ones
that are not clearly defeated by any other candidate. 

Nevertheless, relevant characteristic of this method is that, although it 
is a Condorcet method (if there is a Condorcet winner it is elected winner), 
a Condorcet loser might be also elected winner.
}
