% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_classifications.R
\docType{data}
\name{structural_classifications}
\alias{structural_classifications}
\title{Example structural classifications}
\format{
A tibble containing consensus structural classifications for example \emph{m/z} features.
}
\source{
These example consensus structural classifications are based on the \emph{Brachypodium distachyon} ecotype
comparison FIE-MS metabolomic fingerprinting data set available in the
\href{https://aberhrml.github.io/metaboData/}{{metaboData}} package.
}
\usage{
structural_classifications
}
\description{
Example structural classifications
}
\keyword{datasets}
