% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{tapproval}
\alias{tapproval}
\title{t-approval ranking rule}
\usage{
tapproval(profileOfRankings, t = 2, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
t-approval ranking rule
}
\seealso{
Other ranking rule: 
\code{\link{borda_count}()},
\code{\link{borda_winner}()},
\code{\link{scoring}()},
\code{\link{veto}()}
}
\concept{ranking rule}
