% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{threshold}
\alias{threshold}
\alias{threshold,Consensus-method}
\title{threshold}
\usage{
threshold(x)

\S4method{threshold}{Consensus}(x)
}
\arguments{
\item{x}{S4 object of class Consensus}
}
\description{
Get the threshold of a Consensus object.
}
