% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{toLatex.por}
\alias{toLatex.por}
\title{Latex table of profile of rankings}
\usage{
\method{toLatex}{por}(profileOfRankings)
}
\arguments{
\item{profileOfRankings}{}
}
\description{
Latex table of profile of rankings
}
