% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{update_candidates}
\alias{update_candidates}
\title{Change the names of the candidates of a given profile}
\usage{
update_candidates(profileOfRankings, newCandidates)
}
\arguments{
\item{profileOfRankings}{Profile to update the name of the candidates.}

\item{newCandidates}{New name of the candidates.}
}
\value{
The profile of rankings with the name of the candidates modified.
}
\description{
Change the names of the candidates of a given profile
}
