% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{veto}
\alias{veto}
\title{Veto (a.k.a. antiplurality) ranking rule}
\usage{
veto(profileOfRankings, seeTrace = FALSE, seePoints = FALSE)
}
\arguments{
\item{seeTrace}{}
}
\value{

}
\description{
Veto (a.k.a. antiplurality) ranking rule
}
\seealso{
Other ranking rule: 
\code{\link{borda_winner}()},
\code{\link{borda}()},
\code{\link{scoring_rule}()},
\code{\link{tapproval}()}
}
\concept{ranking rule}
