% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_wvm.R
\name{wvm}
\alias{wvm}
\title{Weigted Voting Method (WVM)}
\usage{
wvm(profileOfRankings, alpha = 0.5, seePoints = FALSE)
}
\arguments{
\item{profileOfRankings}{}
}
\value{

}
\description{
Ranking rule
}
\details{
This procedure is based on the proportion of victories/defeats of each couple 
of candidates. It candidate receives a score resulting of the aggregation of 
all the proportions of victories and ranks the candidates according to this score.
We consider the matrix \eqn{P} of proportions of victories/defeats where \eqn{P_{ij}} is 
the proportion of times that candidate \eqn{c_i} is preferred to candidate \eqn{c_j}{cj}, 
for any \eqn{i != j}{ij}, and \eqn{P_{ii} = 0.5}. The victories and the defeats are separated 
in two matrices \eqn{P^+}  and \eqn{P^−}

t must be remarked that this method comes from the field of group decision making [33].
}
