% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R, R/construction.R
\name{mf}
\alias{mf}
\alias{mf,Consensus-method}
\alias{adductRules}
\alias{adductRules,Consensus-method}
\alias{organism}
\alias{organism,Consensus-method}
\alias{database}
\alias{database,Consensus-method}
\alias{hits}
\alias{hits,Consensus-method}
\alias{PIPs}
\alias{PIPs,Consensus-method}
\alias{classifications}
\alias{classifications,Consensus-method}
\alias{consensusClassifications}
\alias{consensusClassifications,Consensus-method}
\alias{consensus}
\alias{consensus,Consensus-method}
\alias{classifications,Construction-method}
\alias{summariseClassifications}
\alias{summariseClassifications,Construction-method}
\title{Accessor methods for the \code{Consensus} and \code{Construction} S4 classes}
\usage{
mf(x)

\S4method{mf}{Consensus}(x)

adductRules(x)

\S4method{adductRules}{Consensus}(x)

organism(x)

\S4method{organism}{Consensus}(x)

database(x)

\S4method{database}{Consensus}(x)

hits(x)

\S4method{hits}{Consensus}(x)

PIPs(x)

\S4method{PIPs}{Consensus}(x)

classifications(x)

\S4method{classifications}{Consensus}(x)

consensusClassifications(x)

\S4method{consensusClassifications}{Consensus}(x)

consensus(x, adduct, threshold = 50)

\S4method{consensus}{Consensus}(x, adduct, threshold = 50)

\S4method{classifications}{Construction}(x)

summariseClassifications(x)

\S4method{summariseClassifications}{Construction}(x)
}
\arguments{
\item{x}{object of S4  class \code{Consensus} or \code{Construction}}

\item{adduct}{the ionisation adduct for which a consensus should be calculated}

\item{threshold}{the percentage majority threshold for consensus classification}
}
\value{
A character, a numeric, a tibble or an object of S4 class \code{MetaboliteDatabase}, depending on the method used.
}
\description{
Accessor methods for the \code{Consensus} and \code{Construction} S4 classes.
}
\details{
\itemize{
\item \code{mf} - Return the searched molecular formula
\item \code{adductRules} - Return a tibble of adduct formation rules.
\item \code{organism} - Return the KEGG organism ID.
\item \code{database} - Return the searched database.
\item \code{threshold} - Return the percentage consensus threshold for structural classification selection.
\item \code{hits} - Return a \code{MetaboliteDatabase} ionisation database of matched database hits.
\item \code{PIPs} - Return the putative ionisation products of database hits
\item \code{classifications} -Return the structural chemical classifications of database hits.
\item \code{consensusClassifications} - Return the consensus classification or classifications.
\item \code{summariseClassifications} - Return a tibble of summarised consensus structural classifications.
}
}
\examples{
consensus <- construct(
  'C4H6O5',
  organism = 'hsa')

## Return the molecular formula
mf(consensus)

## Return the adduct formation rules
adductRules(consensus)

## Return the KEGG organism ID
organism(consensus)

## Return the searched database
database(consensus)

## Return the `MetaboliteDatabase` ionisation database of searched database  hits
hits(consensus)

## Return the putative ionisation products
PIPs(consensus)

## Return the structural classifications
classifications(consensus)

## Return the consensus structural classification
consensusClassifications(consensus)
}
