% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHV.R
\name{BHVDistance}
\alias{BHVDistance}
\alias{BHVDist}
\alias{BHV}
\title{Geodesic (BHV) distance between two trees}
\usage{
BHVDistance(tree1, tree2)

BHVDist(tree1, tree2)

BHV(tree1, tree2)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, sharing the same leaf labels and
carrying \code{edge.length}.}
}
\value{
\code{BHVDistance()} returns a single non-negative number: the length of
the geodesic between \code{tree1} and \code{tree2} in BHV treespace.
}
\description{
\code{BHVDistance()} returns the geodesic distance between two phylogenetic trees
with edge lengths in the Billera-Holmes-Vogtmann (BHV) treespace
\insertCite{BilleraHolmesVogtmann2001}{ConsTree}, computed with the
polynomial-time GTP algorithm of \insertCite{OwenProvan2011;textual}{ConsTree}.
}
\examples{
set.seed(2)
t1 <- TreeTools::RandomTree(8, root = TRUE, lengths = runif(14))
t2 <- TreeTools::RandomTree(8, root = TRUE, lengths = runif(14))
BHVDistance(t1, t2)

}
\references{
\insertAllCited{}
}
\seealso{
Other BHV summaries:
\code{\link[=BHVMean]{BHVMean()}},
\code{\link[=BHVPairwiseDistances]{BHVPairwiseDistances()}}
}
\concept{BHV summaries}
