% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ranking_rule.R,
%   R/scoring_ranking_rule_c.R
\name{borda_count}
\alias{borda_count}
\title{Title}
\usage{
borda_count(profileOfRankings, verbose = FALSE, seePoints = FALSE)

borda_count(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{



}
\description{
Title

Title
}
