% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{borda_winner}
\alias{borda_winner}
\title{Borda Winner}
\usage{
borda_winner(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Applies Borda Count in the profile of rankings given as first parameter and
then it takes the winner in the first position.
}
\seealso{
Other ranking rule: 
\code{\link{borda_count}()},
\code{\link{scoring}()},
\code{\link{tapproval}()},
\code{\link{veto}()}
}
\concept{ranking rule}
