% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/consensus.R
\docType{methods}
\name{consensus}
\alias{consensus}
\alias{consensus,Assignment-method}
\title{consensus}
\usage{
consensus(x, organism = "hsa", threshold = 0.5)

\S4method{consensus}{Assignment}(x, organism = "hsa", threshold = 0.5)
}
\arguments{
\item{x}{S4 object of class Workflow}

\item{organism}{organism kegg ID.}

\item{threshold}{majority assignment threshold for consensus classifications}
}
\description{
Consensus classifications for molecular formula assignments.
}
\examples{
library(MFassign) 
p <- assignmentParameters('FIE')
p@nCores <- 2
assignment <- assignMFs(peakData,p)

consensusCl <- consensus(assignment)
}
