% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{Consensus structural classifications for a molecular formula}
\usage{
construct(
  MF,
  db = c("kegg", "pubchem"),
  organism = character(),
  threshold = 50,
  adduct_rules_table = adduct_rules(),
  classyfireR_cache = NULL
)
}
\arguments{
\item{MF}{the molecular formula to search}

\item{db}{databases to search. Can be either \code{kegg} or \code{pubchem}.}

\item{organism}{the KEGG organism ID to search. Ignored if argument \code{db} includes \code{pubchem}.}

\item{threshold}{the percentage majority threshold for consensus classification}

\item{adduct_rules_table}{a dataframe containing adduct formation rules. The defaults is \code{mzAnnotation::adduct_rules()}.}

\item{classyfireR_cache}{file path for a \code{classyfireR} cache. See the documentation of \code{classyfireR::get_classification} for more details.}
}
\value{
An S4 object of class \code{Consensus}.
}
\description{
Generate a consensus structural classification for a molecular formula.
}
\examples{
construct('C4H6O5')
}
