% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{construct}
\usage{
construct(MF, db = c("kegg", "pubchem"), organism = character(),
  adductRules = adducts(), threshold = 50)
}
\arguments{
\item{MF}{molecular formula}

\item{db}{databases to search. Can be either kegg or pubchem.}

\item{organism}{KEGG organism ID. Ignored if kegg is not specified in db.}

\item{adductRules}{adduct rules table as returned by mzAnnotation::adducts()}

\item{threshold}{% majority threshold for consensus classifications}
}
\description{
Construct consensus structural classifications for a given molecular formula.
}
