% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{construction}
\alias{construction}
\alias{construction,tbl_df-method}
\alias{construction,Assignment-method}
\title{construction}
\usage{
construction(
  x,
  library_path = tempdir(),
  db = c("kegg", "pubchem"),
  organism = character(),
  threshold = 50,
  adduct_rules_table = adduct_rules(),
  classyfireR_cache = NULL
)

\S4method{construction}{tbl_df}(
  x,
  library_path = tempdir(),
  db = c("kegg", "pubchem"),
  organism = character(),
  threshold = 50,
  adduct_rules_table = adduct_rules(),
  classyfireR_cache = NULL
)

\S4method{construction}{Assignment}(
  x,
  library_path = tempdir(),
  db = c("kegg", "pubchem"),
  organism = character(),
  threshold = 50,
  adduct_rules_table = adduct_rules(),
  classyfireR_cache = NULL
)
}
\arguments{
\item{x}{Molecular formulas and adducts to search. Should be a tibble containing two character columns named MF and Adduct.}

\item{library_path}{target file library_path for classification library for storing consensus classifications}

\item{db}{databases to search. Can be either kegg or pubchem.}

\item{organism}{KEGG organism ID. Ignored if kegg is not specified in db.}

\item{threshold}{percentage majority threshold for consensus classifications}

\item{adduct_rules_table}{data frame containing adduct formation rules. The defaults is \code{mzAnnotation::adduct_rules()}.}

\item{classyfireR_cache}{file library_path for a \code{classyfireR} cache. See the documentation of \code{classyfireR::get_classification} for more details.}
}
\description{
Build or add to and load a consensus classification library.
}
\examples{
\dontrun{
x <- tibble(MF = c(rep('C12H22O11',2),'C4H6O5'),
              Adduct = c('[M-H]1-','[M+Cl]1-','[M-H]1-'))
structural_classifications <- construction(x)
} 
}
