% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{construction}
\alias{construction}
\title{construction}
\usage{
construction(
  MFs,
  path = ".",
  db = c("kegg", "pubchem"),
  organism = character(),
  threshold = 50
)
}
\arguments{
\item{MFs}{Molecular formulas and adducts to search. Should be a tibble containing two character columns named MF and Adduct.}

\item{path}{target file path for classification library}

\item{db}{databases to search. Can be either kegg or pubchem.}

\item{organism}{KEGG organism ID. Ignored if kegg is not specified in db.}

\item{threshold}{\% majority threshold for consensus classifications}
}
\description{
Build or add to and load a consensus classification library.
}
\examples{
\dontrun{
MFs <- tibble(MF = c(rep('C12H22O11',2),'C4H6O5'),
              Adduct = c('[M-H]1-','[M+Cl]1-','[M-H]1-'))
structural_classifications <- construction(MFs)
} 
}
