% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{keggConsensus}
\alias{keggConsensus}
\title{keggConsensus}
\usage{
keggConsensus(MFs, organism = NULL, adductRules = adducts(),
  threshold = 50)
}
\arguments{
\item{MFs}{tibble containing 2 columns. The first named MF containing molecular formulas. The second named Adduct containing adducts. See example.}

\item{organism}{KEGG organism ID. If NULL use all compounds.}

\item{adductRules}{tibble containing adduct formation rules as returned by mzAnnotation::adducts()}

\item{threshold}{\% majority assignment threshold for consensus classifications}
}
\description{
Return consensus structural classifications for given molecular formulas and their adducts based on the KEGG compound database.
}
\examples{
\dontrun{
d <- tibble(MF = c('C12H22O11','C4H6O5'),Adduct = rep('[M-H]1-',2))
keggConsensus(d,'hsa')
}
}
