% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority.R
\name{majority}
\alias{majority}
\title{Majority}
\usage{
majority(profileOfRankings, method = NULL, c1 = NULL, c2 = NULL)
}
\arguments{
\item{profileOfRankings}{}

\item{method}{}

\item{c1}{}

\item{c2}{}
}
\value{

}
\description{
Majority
}
