% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{plurality}
\alias{plurality}
\title{Plurality ranking rule}
\usage{
plurality(profileOfRankings, seeTrace = FALSE, seePoints = FALSE)
}
\arguments{
\item{seeTrace}{}
}
\value{

}
\description{
Plurality ranking rule
}
