% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ranking_rule.R
\name{plurality}
\alias{plurality}
\title{Plurality (first-past-the-post) ranking rule}
\usage{
plurality(profileOfRankings, verbose = FALSE)
}
\arguments{
\item{profileOfRankings}{The profile of ranking where the ranking rule is applied}
}
\value{
The graphic representation of the ranking obtained after applying
plurality to the profile of rankings received as parameter

# Create a profile of rankings
profileOfRankings <- profile.of.rankings(candidates = c('a','b','c','d'))
profileOfRankings <- add.ranking(profileOfRankings, c('c','b','a','d'), votes = 6)
profileOfRankings <- add.ranking(profileOfRankings, c('a','d','b','c'), votes = 5)
profileOfRankings <- add.ranking(profileOfRankings, c('b','a','d','c'), votes = 3)

# Calculate the ranking
plurality(profileOfRankings)
}
\description{
Candidates are ranked according to the number of voters that
considered this candidate their preferred option. It leads to a ranking rule
where only the main choice of each voter is taken into account.
}
\keyword{plurality}
\keyword{ranking}
\keyword{rule,}
