% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{plurality}
\alias{plurality}
\title{plurality ranking rule}
\usage{
plurality(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
plurality ranking rule
}
