% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{random_profile_of_rankings}
\alias{random_profile_of_rankings}
\title{Create a random profile of rankings}
\usage{
random_profile_of_rankings(
  ncandidates = 4,
  nvoters = 10,
  seed = NULL,
  withties = FALSE,
  cnames = NULL,
  distinct = NULL,
  distribution = "norm"
)
}
\arguments{
\item{ncandidates}{Number of candidates. Default is 4.}

\item{nvoters}{Number of voters. Default is 10.}

\item{seed}{To fix a seed to replicate the results. Default is NULL.}

\item{withties}{Indicates whether the profile created can contain ties. Default is FALSE.}

\item{cnames}{Names of the candidates}

\item{distinct}{Num}

\item{distribution}{norm or uniform}
}
\value{
A profile of rankings
}
\description{
Useful for testing purposes
}
