% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{random_profile_of_rankings}
\alias{random_profile_of_rankings}
\title{Create random profile of rankings}
\usage{
random_profile_of_rankings(
  ncandidates = 4,
  nranking = 10,
  seed = NULL,
  withties = FALSE
)
}
\arguments{
\item{ncandidates}{Number of candidates}

\item{nranking}{Number of rankings in the profile of rankings}

\item{seed}{Fix a seed to hace reproducible code}

\item{withties}{The rankings are generated without ties by default.}
}
\description{
Create random profile of rankings
}
