% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Create ranking from vector}
\usage{
ranking(v, cnames = NULL, decreasing = FALSE)
}
\arguments{
\item{v}{vector to be translated to a ranking}

\item{cnames}{names of the candidates. The values in this vector cannot
must identified uniquely each candidate, thus, repeated values are not 
allowed.}

\item{decreasing}{Default is FALSE.}
}
\description{
Given a numeric vector, the default behaviour creates a ranking in 
ascending value. This is, the lower the value of the vector the better the 
candidate is considered.
If names are given in the vector they will be used as names of the 
candidates, to overwrite them, use the argument `cnames`.
}
\seealso{
Other ranking: 
\code{\link{parse_ranking}()}
}
\concept{ranking}
