% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking_to_linear}
\alias{ranking_to_linear}
\title{Ranking to lineal}
\usage{
ranking_to_linear(ranking)
}
\arguments{
\item{ranking}{}
}
\value{
linear extension of the ranking given as parameter
}
\description{
Take a ranking that contains ties and give a linear extension of the ranking
}
