% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/por.R
\name{read_rankings}
\alias{read_rankings}
\title{Read rankings from file}
\usage{
read_rankings(file_path, from_csv = FALSE)
}
\arguments{
\item{file_path}{Path of the file storing the rankings}

\item{from_csv}{Default FALSE}
}
\description{
Given a .csv file, read the rankings in the file and creates
a profile of rankings object.
}
