% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{reversals}
\alias{reversals}
\title{Calculate all the possible rankings}
\usage{
reversals(candidates, verbose = FALSE)
}
\arguments{
\item{candidates}{}

\item{verbose}{}
}
\value{

}
\description{
Calculate all the possible rankings
}
