% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{scoring}
\alias{scoring}
\title{Scoring ranking rules}
\usage{
scoring(
  profileOfRankings,
  method = NULL,
  t = 1,
  seeTrace = FALSE,
  seePoints = FALSE
)
}
\arguments{
\item{profileOfRankings}{object of the class profile of rankings}

\item{method}{\itemize{
 \item{"plurality"}{Stuff}
 \item{"borda"}{Stuff}
 \item{"t"}{Stuff}
}}

\item{t}{only necessary when the \code{method} choosen is \code{t}}

\item{seeTrace}{by default FALSE. Change to TRUE for seeing on the screen
the workflow of the function}
}
\value{
the ranking generated after applying the ranking rule
}
\description{
Scoring ranking rules
}
\examples{


}
\seealso{
Other ranking rule: 
\code{\link{borda_count}()},
\code{\link{borda_winner}()},
\code{\link{tapproval}()},
\code{\link{veto}()}
}
\concept{ranking rule}
