% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ranking_rule.R
\name{scoring}
\alias{scoring}
\title{Scoring ranking rules}
\usage{
scoring(profileOfRankings, method = NULL, t = 1, verbose = FALSE,
  seePoints = FALSE)
}
\arguments{
\item{profileOfRankings}{object of the class profile of rankings}

\item{method}{\itemize{
 \item{"plurality"}{Stuff}
 \item{"borda"}{Stuff}
 \item{"t"}{Stuff}
}}

\item{t}{only necessary when the \code{method} choosen is}

\item{verbose}{by default FALSE. Change to TRUE for seeing on the screen
the workflow of the function}
}
\value{
the ranking generated after applying the ranking rule
}
\description{
Scoring ranking rules
}
\examples{


}
