% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_scoring.R
\name{veto}
\alias{veto}
\title{veto ranking rule}
\usage{
veto(profileOfRankings, verbose = FALSE, seePoints = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
veto ranking rule
}
\seealso{
Other ranking rule: 
\code{\link{borda_count}()},
\code{\link{borda_winner}()},
\code{\link{scoring}()},
\code{\link{tapproval}()}
}
\concept{ranking rule}
